pyproject.toml
[tool.poetry]
name = "mudmind"
version = "0.1.0"
description = "MudMind MVP - текстовые MUD игры с ИИ ассистентом"
authors = ["GrayAngelof"]
readme = "README.md"
license = "MIT"
python = "^3.12"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "2.3.5"
jinja2 = "3.1.6"
typing-extensions = "4.15.0"
diskcache = "5.6.3"
httpx = "0.28.1"
requests = "2.32.5"
anyio = "4.12.0"
MarkupSafe = "3.0.3"
more-itertools = "10.8.0"
msgpack = "1.1.2"
packaging = "25.0"
setuptools = "80.9.0"
wheel = "0.45.1"
zstandard = "0.25.0"

# Локальная установка llama_cpp_python
llama_cpp_python = {path = "wheels/llama_cpp_python-0.3.8+cu128.gemma3-cp312-cp312-win_amd64.whl"}

[tool.poetry.dev-dependencies]
# Добавьте, если нужны тесты или линтеры
pytest = "^8.3"
black = "^24.9"
